articlesproject/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-admin-cli==0.1.1
django-filter==25.1
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.1
coverage==7.2.7
